Refactor Apple app UI structure
