+++ app/apple/herm/Models/CPSLChatModel.swift
@@ -1,0 +1,228 @@
+import Combine
+import Foundation
+
+@MainActor
+final class CPSLChatModel: ObservableObject {
+    @Published var promptText = ""
+    @Published var comingSoonMessage: String?
+    @Published private(set) var messages: [CPSLChatMessage] = []
+    @Published private(set) var isRunning = false
+    @Published private(set) var isFileBrowserOpen = false
+    @Published private(set) var browserPath = "/"
+    @Published private(set) var browserEntries: [CPSLFileEntry] = []
+    @Published private(set) var childEntriesByPath: [String: [CPSLFileEntry]] = [:]
+    @Published private(set) var expandedFilePaths: Set<String> = []
+    @Published private(set) var loadingFilePaths: Set<String> = []
+    @Published private(set) var fileBrowserError: String?
+
+    private let service = CPSLDebugService()
+
+    func showComingSoon(_ message: String = "coming soon") {
+        comingSoonMessage = message
+    }
+
+    func submitPrompt() {
+        let input = promptText.trimmingCharacters(in: .whitespacesAndNewlines)
+        guard !input.isEmpty, !isRunning else {
+            return
+        }
+
+        promptText = ""
+        isFileBrowserOpen = false
+        if input.hasPrefix("!") {
+            let command = String(input.dropFirst()).trimmingCharacters(in: .whitespacesAndNewlines)
+            guard !command.isEmpty else {
+                appendErrorMessage(title: nil, body: "Enter a command after !")
+                return
+            }
+            runCommand(command)
+            return
+        }
+
+        messages.append(CPSLChatMessage(role: .user, title: nil, body: input))
+    }
+
+    func toggleFileBrowser() {
+        isFileBrowserOpen.toggle()
+
+        if isFileBrowserOpen && browserEntries.isEmpty && loadingFilePaths.isEmpty {
+            loadBrowserPath("/")
+        }
+    }
+
+    func closeFileBrowser() {
+        isFileBrowserOpen = false
+    }
+
+    func loadBrowserPath(_ path: String) {
+        let normalized = normalizedPath(path)
+        browserPath = normalized
+        fileBrowserError = nil
+        expandedFilePaths.removeAll()
+        childEntriesByPath.removeAll()
+
+        Task {
+            await loadDirectory(normalized, childOf: nil)
+        }
+    }
+
+    func navigateToParentDirectory() {
+        loadBrowserPath(parentPath(of: browserPath))
+    }
+
+    func toggleExpansion(for entry: CPSLFileEntry) {
+        guard entry.isDirectory else {
+            return
+        }
+
+        if expandedFilePaths.contains(entry.path) {
+            expandedFilePaths.remove(entry.path)
+            return
+        }
+
+        expandedFilePaths.insert(entry.path)
+        if childEntriesByPath[entry.path] == nil {
+            Task {
+                await loadDirectory(entry.path, childOf: entry.path)
+            }
+        }
+    }
+
+    func openFileEntry(_ entry: CPSLFileEntry) {
+        if entry.isDirectory {
+            loadBrowserPath(entry.path)
+        } else {
+            showComingSoon("coming soon")
+        }
+    }
+
+    func children(for path: String) -> [CPSLFileEntry] {
+        childEntriesByPath[path] ?? []
+    }
+
+    func isExpanded(_ entry: CPSLFileEntry) -> Bool {
+        expandedFilePaths.contains(entry.path)
+    }
+
+    func isLoading(_ path: String) -> Bool {
+        loadingFilePaths.contains(path)
+    }
+
+    private func runCommand(_ command: String) {
+        let message = CPSLChatMessage(role: .command, title: nil, body: commandBlockBody(command: command))
+        messages.append(message)
+        isRunning = true
+
+        Task {
+            let result = await service.evaluate(command)
+            applyCommandResult(result, command: command, messageID: message.id)
+            isRunning = false
+        }
+    }
+
+    private func applyCommandResult(_ result: CPSLEvalServiceResult, command: String, messageID: UUID) {
+        let body = commandBlockBody(command: command, result: result)
+        guard let index = messages.firstIndex(where: { $0.id == messageID }) else {
+            messages.append(CPSLChatMessage(role: .command, title: nil, body: body))
+            return
+        }
+        messages[index].body = body
+    }
+
+    private func commandBlockBody(command: String, result: CPSLEvalServiceResult? = nil) -> String {
+        var sections = ["!\(command)"]
+        guard let result else {
+            return sections.joined(separator: "\n\n")
+        }
+
+        var outputSections: [String] = []
+        outputSections.append(contentsOf: result.warnings.map { "warning: \($0)" })
+        appendTrimmed(result.stdout, to: &outputSections)
+        appendTrimmed(result.stderr, to: &outputSections)
+
+        if let ffiError = result.ffiError {
+            outputSections.append(ffiError)
+        }
+        if let errorMessage = result.errorMessage {
+            let prefix = result.errorCode.map { "error[\($0)]" } ?? "error"
+            outputSections.append("\(prefix): \(errorMessage)")
+        }
+        if result.errorCode == "invalid_response", let rawJSON = result.rawJSON {
+            outputSections.append(rawJSON)
+        }
+        if outputSections.isEmpty {
+            let exit = result.exitCode.map { "exit \($0)" } ?? "done"
+            outputSections.append(exit)
+        }
+
+        sections.append(outputSections.joined(separator: "\n\n"))
+        return sections.joined(separator: "\n\n")
+    }
+
+    private func appendTrimmed(_ text: String, to sections: inout [String]) {
+        let trimmed = text.trimmingCharacters(in: .newlines)
+        guard !trimmed.isEmpty else {
+            return
+        }
+        sections.append(trimmed)
+    }
+
+    private func loadDirectory(_ path: String, childOf parent: String?) async {
+        guard !loadingFilePaths.contains(path) else {
+            return
+        }
+        loadingFilePaths.insert(path)
+        defer {
+            loadingFilePaths.remove(path)
+        }
+
+        let listing = await service.listDirectory(path)
+        if let error = listing.error {
+            applyDirectoryLoadFailure(error, path: path, childOf: parent)
+            return
+        }
+
+        if let parent {
+            childEntriesByPath[parent] = listing.entries
+        } else {
+            browserEntries = listing.entries
+        }
+    }
+
+    private func applyDirectoryLoadFailure(_ message: String, path: String, childOf parent: String?) {
+        if let parent {
+            childEntriesByPath[parent] = []
+        } else {
+            browserEntries = []
+        }
+        fileBrowserError = "\(path): \(message)"
+    }
+
+    private func appendErrorMessage(title: String?, body: String) {
+        messages.append(CPSLChatMessage(role: .error, title: title, body: body))
+    }
+
+    private func normalizedPath(_ path: String) -> String {
+        var normalized = path.isEmpty ? "/" : path
+        if !normalized.hasPrefix("/") {
+            normalized = "/\(normalized)"
+        }
+        while normalized.count > 1 && normalized.hasSuffix("/") {
+            normalized.removeLast()
+        }
+        return normalized
+    }
+
+    private func parentPath(of path: String) -> String {
+        let normalized = normalizedPath(path)
+        guard normalized != "/" else {
+            return "/"
+        }
+
+        let components = normalized.split(separator: "/")
+        guard components.count > 1 else {
+            return "/"
+        }
+        return "/" + components.dropLast().joined(separator: "/")
+    }
+}
