+++ app/apple/herm/Views/Composer/CPSLPromptComposerView.swift
@@ -1,0 +1,264 @@
+import Dispatch
+import SwiftUI
+#if canImport(UIKit)
+import UIKit
+#endif
+
+#if canImport(UIKit)
+private struct CPSLPromptTextView: UIViewRepresentable {
+    @Binding var text: String
+    let isCommandInput: Bool
+    let isDisabled: Bool
+    let maxHeight: CGFloat
+    let dismissKeyboardRequest: Int
+    let onHeightChange: (CGFloat) -> Void
+
+    func makeUIView(context: Context) -> UITextView {
+        let textView = UITextView()
+        textView.delegate = context.coordinator
+        textView.backgroundColor = .clear
+        textView.textColor = UIColor(CPSLTheme.text)
+        textView.tintColor = UIColor(CPSLTheme.text)
+        textView.font = UIFont.systemFont(ofSize: CPSLTheme.normalTextSize, weight: .regular)
+        textView.textContainerInset = .zero
+        textView.textContainer.lineFragmentPadding = 0
+        textView.returnKeyType = .default
+        textView.setContentCompressionResistancePriority(.defaultLow, for: .horizontal)
+        context.coordinator.dismissKeyboardRequest = dismissKeyboardRequest
+        applyInputTraits(to: textView)
+        return textView
+    }
+
+    func updateUIView(_ textView: UITextView, context: Context) {
+        context.coordinator.parent = self
+
+        if textView.text != text {
+            textView.text = text
+        }
+
+        let didChangeTraits = applyInputTraits(to: textView)
+        textView.isEditable = !isDisabled
+        textView.isSelectable = !isDisabled
+        textView.isScrollEnabled = textView.contentSize.height > maxHeight
+
+        if isDisabled && textView.isFirstResponder {
+            textView.resignFirstResponder()
+        } else if context.coordinator.dismissKeyboardRequest != dismissKeyboardRequest {
+            context.coordinator.dismissKeyboardRequest = dismissKeyboardRequest
+            textView.resignFirstResponder()
+        }
+
+        if didChangeTraits && textView.isFirstResponder {
+            textView.reloadInputViews()
+        }
+
+        context.coordinator.reportHeight(for: textView)
+    }
+
+    func makeCoordinator() -> Coordinator {
+        Coordinator(parent: self)
+    }
+
+    @discardableResult
+    private func applyInputTraits(to textView: UITextView) -> Bool {
+        let keyboardType: UIKeyboardType = isCommandInput ? .asciiCapable : .default
+        let autocapitalizationType: UITextAutocapitalizationType = isCommandInput ? .none : .sentences
+        let autocorrectionType: UITextAutocorrectionType = isCommandInput ? .no : .yes
+        let spellCheckingType: UITextSpellCheckingType = isCommandInput ? .no : .default
+        let smartQuotesType: UITextSmartQuotesType = isCommandInput ? .no : .default
+        let smartDashesType: UITextSmartDashesType = isCommandInput ? .no : .default
+        let smartInsertDeleteType: UITextSmartInsertDeleteType = isCommandInput ? .no : .default
+
+        let didChange = textView.keyboardType != keyboardType ||
+            textView.autocapitalizationType != autocapitalizationType ||
+            textView.autocorrectionType != autocorrectionType ||
+            textView.spellCheckingType != spellCheckingType ||
+            textView.smartQuotesType != smartQuotesType ||
+            textView.smartDashesType != smartDashesType ||
+            textView.smartInsertDeleteType != smartInsertDeleteType
+
+        textView.keyboardType = keyboardType
+        textView.autocapitalizationType = autocapitalizationType
+        textView.autocorrectionType = autocorrectionType
+        textView.spellCheckingType = spellCheckingType
+        textView.smartQuotesType = smartQuotesType
+        textView.smartDashesType = smartDashesType
+        textView.smartInsertDeleteType = smartInsertDeleteType
+        return didChange
+    }
+
+    final class Coordinator: NSObject, UITextViewDelegate {
+        var parent: CPSLPromptTextView
+        var dismissKeyboardRequest = 0
+
+        init(parent: CPSLPromptTextView) {
+            self.parent = parent
+        }
+
+        func textViewDidChange(_ textView: UITextView) {
+            parent.text = textView.text
+            reportHeight(for: textView)
+        }
+
+        func reportHeight(for textView: UITextView) {
+            let fittingSize = CGSize(width: textView.bounds.width, height: .greatestFiniteMagnitude)
+            let height = textView.sizeThatFits(fittingSize).height
+            DispatchQueue.main.async {
+                self.parent.onHeightChange(height)
+            }
+        }
+    }
+}
+#endif
+
+struct CPSLPromptComposerView: View {
+    @ObservedObject var model: CPSLChatModel
+    let dismissKeyboardRequest: Int
+    let dismissKeyboard: () -> Void
+    @State private var promptContentHeight: CGFloat = 0
+
+    private var hasPromptInput: Bool {
+        !model.promptText.trimmingCharacters(in: .whitespacesAndNewlines).isEmpty
+    }
+
+    private var isCommandInput: Bool {
+        model.promptText.trimmingCharacters(in: .whitespacesAndNewlines).hasPrefix("!")
+    }
+
+    private var promptLineHeight: CGFloat {
+#if canImport(UIKit)
+        ceil(UIFont.systemFont(ofSize: CPSLTheme.normalTextSize, weight: .regular).lineHeight)
+#else
+        ceil(CPSLTheme.normalTextSize * 1.25)
+#endif
+    }
+
+    private var promptTextHeight: CGFloat {
+        min(max(promptContentHeight, promptLineHeight), promptLineHeight * 6)
+    }
+
+    var body: some View {
+        VStack(alignment: .leading, spacing: CPSLTheme.medium) {
+            ZStack(alignment: .topLeading) {
+                if model.promptText.isEmpty {
+                    Text("Ask Anything")
+                        .font(CPSLTheme.bodyFont)
+                        .foregroundStyle(CPSLTheme.mutedText)
+                        .padding(.horizontal, CPSLTheme.medium)
+                        .padding(.vertical, CPSLTheme.small)
+                }
+
+#if canImport(UIKit)
+                CPSLPromptTextView(
+                    text: $model.promptText,
+                    isCommandInput: isCommandInput,
+                    isDisabled: model.isRunning,
+                    maxHeight: promptLineHeight * 6,
+                    dismissKeyboardRequest: dismissKeyboardRequest
+                ) { height in
+                    promptContentHeight = height
+                }
+                .frame(height: promptTextHeight)
+                .padding(.horizontal, CPSLTheme.medium)
+                .padding(.vertical, CPSLTheme.small)
+#else
+                TextField("", text: $model.promptText, axis: .vertical)
+                    .textFieldStyle(.plain)
+                    .submitLabel(.return)
+                    .lineLimit(1...6)
+                    .font(CPSLTheme.bodyFont)
+                    .foregroundStyle(CPSLTheme.text)
+                    .tint(CPSLTheme.text)
+                    .disabled(model.isRunning)
+                    .padding(.horizontal, CPSLTheme.medium)
+                    .padding(.vertical, CPSLTheme.small)
+#endif
+            }
+            .background {
+                RoundedRectangle(cornerRadius: CPSLTheme.controlRadius, style: .continuous)
+                    .fill(isCommandInput ? CPSLTheme.command.opacity(0.82) : Color.clear)
+            }
+            .animation(.easeOut(duration: 0.16), value: isCommandInput)
+
+            HStack(spacing: CPSLTheme.medium) {
+                Button {
+                    dismissKeyboard()
+                    model.showComingSoon("coming soon")
+                } label: {
+                    Image(systemName: "plus")
+                        .font(.system(size: 18, weight: .medium))
+                        .frame(width: CPSLTheme.controlSize, height: CPSLTheme.controlSize)
+                        .contentShape(Rectangle())
+                }
+                .buttonStyle(.plain)
+                .foregroundStyle(CPSLTheme.text)
+                .cpslGlassBackground(
+                    in: RoundedRectangle(cornerRadius: CPSLTheme.controlRadius, style: .continuous),
+                    tint: CPSLTheme.card.opacity(0.38),
+                    strokeOpacity: 0.045
+                )
+                .contentShape(RoundedRectangle(cornerRadius: CPSLTheme.controlRadius, style: .continuous))
+
+                Spacer()
+
+                Button {
+                    dismissKeyboard()
+                    model.showComingSoon("coming soon")
+                } label: {
+                    Image(systemName: "mic.fill")
+                        .font(.system(size: 15, weight: .medium))
+                        .frame(width: CPSLTheme.controlSize, height: CPSLTheme.controlSize)
+                        .contentShape(Rectangle())
+                }
+                .buttonStyle(.plain)
+                .foregroundStyle(CPSLTheme.text)
+                .cpslGlassBackground(
+                    in: RoundedRectangle(cornerRadius: CPSLTheme.controlRadius, style: .continuous),
+                    tint: CPSLTheme.card.opacity(0.38),
+                    strokeOpacity: 0.045
+                )
+                .contentShape(RoundedRectangle(cornerRadius: CPSLTheme.controlRadius, style: .continuous))
+
+                Button {
+                    dismissKeyboard()
+                    if hasPromptInput {
+                        model.submitPrompt()
+                    } else {
+                        model.showComingSoon("coming soon")
+                    }
+                } label: {
+                    Group {
+                        if hasPromptInput {
+                            Image(systemName: "arrow.up")
+                                .font(.system(size: 17, weight: .semibold))
+                                .frame(width: CPSLTheme.controlSize, height: CPSLTheme.controlSize)
+                        } else {
+                            HStack(spacing: CPSLTheme.small) {
+                                Image(systemName: "waveform")
+                                    .font(.system(size: 15, weight: .medium))
+                                Text("Speak")
+                                    .font(CPSLTheme.controlFont)
+                            }
+                            .padding(.horizontal, CPSLTheme.medium)
+                            .frame(height: CPSLTheme.controlSize)
+                        }
+                    }
+                    .contentShape(RoundedRectangle(cornerRadius: CPSLTheme.controlRadius, style: .continuous))
+                }
+                .buttonStyle(.plain)
+                .foregroundStyle(CPSLTheme.background)
+                .background(CPSLTheme.text)
+                .clipShape(RoundedRectangle(cornerRadius: CPSLTheme.controlRadius, style: .continuous))
+                .contentShape(RoundedRectangle(cornerRadius: CPSLTheme.controlRadius, style: .continuous))
+            }
+        }
+        .padding(CPSLTheme.medium)
+        .cpslGlassBackground(
+            in: RoundedRectangle(cornerRadius: CPSLTheme.composerRadius, style: .continuous),
+            tint: CPSLTheme.background.opacity(0.54),
+            strokeOpacity: 0.055
+        )
+        .padding(.horizontal, CPSLTheme.chromeHorizontalInset)
+        .padding(.bottom, CPSLTheme.large)
+    }
+}
